+++ tests/rust.rs
@@ -97,7 +97,7 @@ fn prohibit_duplicate_value_btreeset() {
 
 #[test]
 fn prohibit_duplicate_value_hashmap() {
-    use std::{collections::HashMap, iter::FromIterator};
+    use std::collections::HashMap;
     #[derive(Debug, Eq, PartialEq, Deserialize)]
     struct Doc {
         #[serde(with = "::serde_with::rust::maps_duplicate_key_is_error")]
@@ -163,4 +163,82 @@ fn prohibit_duplicate_value_btreemap() {
     let s = r#"{"map": {"1": 1, "2": 2, "1": 3}}"#;
     let res: Result<Doc, _> = serde_json::from_str(s);
     assert!(res.is_err());
+}
+
+#[test]
+fn duplicate_key_first_wins_hashmap() {
+    use std::collections::HashMap;
+    #[derive(Debug, Eq, PartialEq, Deserialize)]
+    struct Doc {
+        #[serde(with = "::serde_with::rust::maps_first_key_wins")]
+        map: HashMap<usize, usize>,
+    }
+
+    // Different value and key always works
+    let s = r#"{"map": {"1": 1, "2": 2, "3": 3}}"#;
+    let mut v = Doc {
+        map: HashMap::new(),
+    };
+    v.map.insert(1, 1);
+    v.map.insert(2, 2);
+    v.map.insert(3, 3);
+    assert_eq!(v, serde_json::from_str(s).unwrap());
+
+    // Same value for different keys is ok
+    let s = r#"{"map": {"1": 1, "2": 1, "3": 1}}"#;
+    let mut v = Doc {
+        map: HashMap::new(),
+    };
+    v.map.insert(1, 1);
+    v.map.insert(2, 1);
+    v.map.insert(3, 1);
+    assert_eq!(v, serde_json::from_str(s).unwrap());
+
+    // Duplicate keys, the first one is used
+    let s = r#"{"map": {"1": 1, "2": 2, "1": 3}}"#;
+    let mut v = Doc {
+        map: HashMap::new(),
+    };
+    v.map.insert(1, 1);
+    v.map.insert(2, 2);
+    assert_eq!(v, serde_json::from_str(s).unwrap());
+}
+
+#[test]
+fn duplicate_key_first_wins_btreemap() {
+    use std::collections::BTreeMap;
+    #[derive(Debug, Eq, PartialEq, Deserialize)]
+    struct Doc {
+        #[serde(with = "::serde_with::rust::maps_first_key_wins")]
+        map: BTreeMap<usize, usize>,
+    }
+
+    // Different value and key always works
+    let s = r#"{"map": {"1": 1, "2": 2, "3": 3}}"#;
+    let mut v = Doc {
+        map: BTreeMap::new(),
+    };
+    v.map.insert(1, 1);
+    v.map.insert(2, 2);
+    v.map.insert(3, 3);
+    assert_eq!(v, serde_json::from_str(s).unwrap());
+
+    // Same value for different keys is ok
+    let s = r#"{"map": {"1": 1, "2": 1, "3": 1}}"#;
+    let mut v = Doc {
+        map: BTreeMap::new(),
+    };
+    v.map.insert(1, 1);
+    v.map.insert(2, 1);
+    v.map.insert(3, 1);
+    assert_eq!(v, serde_json::from_str(s).unwrap());
+
+    // Duplicate keys, the first one is used
+    let s = r#"{"map": {"1": 1, "2": 2, "1": 3}}"#;
+    let mut v = Doc {
+        map: BTreeMap::new(),
+    };
+    v.map.insert(1, 1);
+    v.map.insert(2, 2);
+    assert_eq!(v, serde_json::from_str(s).unwrap());
 }